refactor: split decorators and utils
