+++ src/common/decorated-loggers.ts
@@ -1,0 +1,1 @@
+export const decoratedLoggers = new Set<string>();
