+++ src/common/logger.decorator.ts
@@ -1,0 +1,9 @@
+import { Inject } from '@nestjs/common';
+
+import { decoratedLoggers } from './decorated-loggers';
+import { getLoggerToken } from './logger.utils';
+
+export function InjectPinoLogger(context = '') {
+  decoratedLoggers.add(context);
+  return Inject(getLoggerToken(context));
+}
