+++ src/constants.ts
@@ -1,3 +1,3 @@
-export const LOGGER_OPTIONS = 'pino-params';
+export const LOGGER_OPTIONS = 'PINO_LOGGER_OPTIONS';
 
 export const LOGGER_KEY = 'logger';
