+++ src/logger-core.module.ts
@@ -1,7 +1,3 @@
-import * as express from 'express';
-import { middleware as ctxMiddleware, setValue } from 'express-ctx';
-import * as pinoHttp from 'pino-http';
-
 import {
   DynamicModule,
   Global,
@@ -13,9 +9,10 @@ import {
 } from '@nestjs/common';
 import { Provider } from '@nestjs/common/interfaces';
 
-import { LOGGER_KEY, LOGGER_OPTIONS } from './constants';
-import { createProvidersForDecorated } from './InjectPinoLogger';
+import { createLoggerMiddlewares } from './common';
+import { LOGGER_OPTIONS } from './constants';
 import { LoggerModuleAsyncOptions, LoggerOptions } from './interfaces';
+import { createProvidersForDecorated } from './logger.providers';
 import { Logger, PinoLogger } from './services';
 
 const DEFAULT_ROUTES = [{ path: '*', method: RequestMethod.ALL }];
@@ -23,10 +20,10 @@ const DEFAULT_ROUTES = [{ path: '*', method: RequestMethod.ALL }];
 @Global()
 @Module({ providers: [Logger], exports: [Logger] })
 export class LoggerCoreModule implements NestModule {
-  static forRoot(params: LoggerOptions | undefined): DynamicModule {
+  static forRoot(options: LoggerOptions | undefined): DynamicModule {
     const paramsProvider: Provider<LoggerOptions> = {
       provide: LOGGER_OPTIONS,
-      useValue: params || {},
+      useValue: options || {},
     };
 
     const decorated = createProvidersForDecorated();
@@ -38,11 +35,11 @@ export class LoggerCoreModule implements NestModule {
     };
   }
 
-  static forRootAsync(params: LoggerModuleAsyncOptions): DynamicModule {
+  static forRootAsync(options: LoggerModuleAsyncOptions): DynamicModule {
     const paramsProvider: Provider<LoggerOptions | Promise<LoggerOptions>> = {
       provide: LOGGER_OPTIONS,
-      useFactory: params.useFactory,
-      inject: params.inject,
+      useFactory: options.useFactory,
+      inject: options.inject,
     };
 
     const decorated = createProvidersForDecorated();
@@ -52,28 +49,30 @@ export class LoggerCoreModule implements NestModule {
       ...decorated,
       PinoLogger,
       paramsProvider,
-      ...(params.providers || []),
+      ...(options.providers || []),
     ];
 
     return {
       module: LoggerCoreModule,
-      imports: params.imports,
+      imports: options.imports,
       providers,
       exports: [Logger, ...decorated, PinoLogger],
     };
   }
 
-  constructor(@Inject(LOGGER_OPTIONS) private readonly params: LoggerOptions) {}
+  constructor(
+    @Inject(LOGGER_OPTIONS) private readonly options: LoggerOptions,
+  ) {}
 
   configure(consumer: MiddlewareConsumer) {
     const {
       exclude,
       forRoutes = DEFAULT_ROUTES,
-      pinoHttp,
+      pinoHttp: pHttp,
       useExisting,
-    } = this.params;
+    } = this.options;
 
-    const middlewares = createLoggerMiddlewares(pinoHttp || {}, useExisting);
+    const middlewares = createLoggerMiddlewares(pHttp || {}, useExisting);
 
     if (exclude) {
       consumer
@@ -84,30 +83,4 @@ export class LoggerCoreModule implements NestModule {
       consumer.apply(...middlewares).forRoutes(...forRoutes);
     }
   }
-}
-
-function createLoggerMiddlewares(
-  params: NonNullable<LoggerOptions['pinoHttp']>,
-  useExisting?: true,
-) {
-  if (useExisting) {
-    return [ctxMiddleware, bindLoggerMiddleware];
-  }
-
-  if (Array.isArray(params)) {
-    return [ctxMiddleware, pinoHttp(...params), bindLoggerMiddleware];
-  }
-
-  // FIXME: params type here is pinoHttp.Options | pino.DestinationStream
-  // pinoHttp has two overloads, each of them takes those types
-  return [ctxMiddleware, pinoHttp(params as any), bindLoggerMiddleware];
-}
-
-function bindLoggerMiddleware(
-  req: express.Request,
-  _res: express.Response,
-  next: express.NextFunction,
-) {
-  setValue(LOGGER_KEY, req.log);
-  next();
 }
