+++ src/logger.module.ts
@@ -5,17 +5,17 @@ import { LoggerCoreModule } from './logger-core.module';
 
 @Module({})
 export class LoggerModule {
-  static forRoot(params?: LoggerOptions | undefined): DynamicModule {
+  static forRoot(options?: LoggerOptions | undefined): DynamicModule {
     return {
       module: LoggerModule,
-      imports: [LoggerCoreModule.forRoot(params)],
+      imports: [LoggerCoreModule.forRoot(options)],
     };
   }
 
-  static forRootAsync(params: LoggerModuleAsyncOptions): DynamicModule {
+  static forRootAsync(options: LoggerModuleAsyncOptions): DynamicModule {
     return {
       module: LoggerModule,
-      imports: [LoggerCoreModule.forRootAsync(params)],
+      imports: [LoggerCoreModule.forRootAsync(options)],
     };
   }
 }
